services/file-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.12
PyPDF2==3.0.1
python-pptx==1.0.2
